PILOT-84 | POC for @extract_args #3
