+++ app/api/base.py
@@ -3,7 +3,6 @@
 """base api"""
 
 from functools import partial, wraps
-import re
 from datetime import datetime, timedelta
 
 import inflection
@@ -16,7 +15,6 @@ from flask_classy import FlaskView
 
 from ..extensions import jwt, auth_datastore
 from ..exceptions import ApplicationException
-from ..utils import extract_dict
 from .errors import api_exception_handler
 from .decorators import token_auth_required, roles_required
 from .representations.json import output_json
@@ -111,51 +109,6 @@ def jwt_error(ex):
 #: make_empty_response(201)
 #: make_empty_response(201, {'header-key': 'value'})
 make_empty_response = partial(make_response, '')
-
-SUPPORTED_OPS = frozenset(['eq', 'ne', 'lt', 'le', 'gt', 'ge', 'lk', 'nl', 'in', 'ni', 'ct', 'mc'])
-
-# sql alchemy ops mapper
-SA_OPS_MAPPER = {
-    'eq': 'eq',
-    'ne': 'ne',
-    'lt': 'lt',
-    'le': 'le',
-    'gt': 'gt',
-    'ge': 'ge',
-    'lk': 'like',
-    'nl': 'notlike',
-    'in': 'in_',
-    'ni': 'notin_',
-    'ct': 'contains',
-    'mc': 'match'
-}
-
-FILTER_KEY_RE = re.compile(r'^(?P<key>\w*[a-zA-Z0-9])+__(?P<op>%s)$' % '|'.join(SUPPORTED_OPS))
-
-
-def extract_filters(args):
-    """Extracts filters then return filters and new args without filter keys"""
-    filters = []
-    # parse filters ?field__op=value => convert to [{key: field, op: op, value:value}, ]
-    filter_dict = extract_dict(args, func=lambda k, v: FILTER_KEY_RE.match(k))
-
-    for key, value in filter_dict.iteritems():
-        matcher = FILTER_KEY_RE.match(key)
-        key = matcher.group('key')
-        operator = matcher.group('op')
-
-        if operator == 'in' or operator == 'ni':
-            value = value.split(',')
-
-        filters.append({
-            'key': key,
-            'op': SA_OPS_MAPPER.get(operator),
-            'value': value
-        })
-
-    args = extract_dict(args, func=lambda k, v: not FILTER_KEY_RE.match(k))
-
-    return filters, args
 
 
 class Resource(FlaskView):
