+++ app/api/decorators.py
@@ -7,6 +7,7 @@ from functools import wraps
 from flask import current_app, _request_ctx_stack
 from flask_principal import Identity, identity_changed
 from flask_jwt import current_user, verify_jwt
+from webargs.flaskparser import use_args
 
 from ..auth import (permissions_required as auth_permissions_required,
                     permissions_accepted as auth_permissions_accepted,
@@ -16,6 +17,7 @@ from ..exceptions import ApplicationException, UnauthorizedException, BadRequest
 from ..pagination import OffsetPagination
 from . import (authenticated, token_authenticated, http_authenticated,
                session_authenticated)
+from .utils import extract_filters
 
 
 def anonymous_required(func):
@@ -299,3 +301,19 @@ def paginated(func):
         }
 
     return decorated
+
+
+def extract_args(arg_map, req=None, locations=None, as_kwargs=False, validate=None):
+
+    def wrapper(func):
+
+        @wraps(func)
+        @use_args(arg_map, req=req, locations=locations, as_kwargs=as_kwargs, validate=validate)
+        def decorated(resource, req_args, *args, **kwargs):
+            filters, req_args = extract_filters(req_args)
+            req_args['filters'] = filters
+
+            return func(resource, req_args, *args, **kwargs)
+
+        return decorated
+    return wrapper
