+++ app/api_1_0/users.py
@@ -6,7 +6,7 @@ from webargs.flaskparser import use_args
 
 from ..auth import admin_role_permission, user_permission
 from ..api import (Resource, marshal_with, marshal_with_data_envelope, token_auth_required, one_of,
-                   anonymous_required, permissions_required, validators, paginated, extract_filters)
+                   anonymous_required, permissions_required, validators, paginated, extract_args)
 from ..api.args import BoolArg
 from ..extensions import auth_datastore
 from ..exceptions import UnauthorizedException
@@ -46,10 +46,8 @@ class UserResource(Resource):
     @permissions_required(admin_role_permission)
     @marshal_with(_user_list_schema)
     @paginated
-    @use_args(search_args)  # TODO(hoate): add support for @extract_args to avoid boilerplate
+    @extract_args(search_args)  # TODO(hoate): add support for @extract_args to avoid boilerplate
     def index(self, args):
-        filters, args = extract_filters(args)
-        args['filters'] = filters
         return auth_datastore.find_users(**args), args
